pyproject.toml
[tool.poetry]
name = "chatbotufs"
version = "0.1.0"
description = "Chatbot básico com RAG usando Gemini (E talvez modelo local)"
authors = ["Caio caiovsa2@gmail.com"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
chromadb = "^0.5.23"
google-genai = "^1.60.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
